Convert all types from int to long

+++ src/main/java/net/projecteuler/problems/Problem009.java
@@ -2,16 +2,16 @@ package net.projecteuler.problems;
 
 public class Problem009 {
 
-  public int productOfPythagoreanTripletWhichSumIsEqualTo(int sumOfPythagoreanTriplet) {
-    int v = 1;
-    int u = 2;
-    int x = 0, y = 0, z = 0;
-    int sum = 0;
+  public long productOfPythagoreanTripletWhichSumIsEqualTo(int sumOfPythagoreanTriplet) {
+    long v = 1;
+    long u = 2;
+    long x = 0, y = 0, z = 0;
+    long sum = 0;
 
     while (sum != sumOfPythagoreanTriplet) {
-      x = (int) (Math.pow(u, 2) - Math.pow(v, 2));
+      x = (long) (Math.pow(u, 2) - Math.pow(v, 2));
       y = 2 * u * v;
-      z = (int) (Math.pow(u, 2) + Math.pow(v, 2));
+      z = (long) (Math.pow(u, 2) + Math.pow(v, 2));
       sum = x + y + z;
       if (sum > sumOfPythagoreanTriplet) {
         v++;
